backend/requirements.txt
fastapi
uvicorn
groq
pydantic
pydantic-settings
python-dotenv
sentence-transformers
numpy
faiss-cpu
pytest
httpx
